(PC-17985) separate homepage fixtures to improve file readability
